pull out CombineDisjunctionToIn tests
