+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/OptimizerRulesTests.java
@@ -86,112 +86,6 @@ public class OptimizerRulesTests extends ESTestCase {
     private static final Expression DUMMY_EXPRESSION =
         new org.elasticsearch.xpack.esql.core.optimizer.OptimizerRulesTests.DummyBooleanExpression(EMPTY, 0);
 
-    //
-    // Constant folding
-    //
-
-    //
-    // CombineDisjunction in Equals
-    //
-    public void testTwoEqualsWithOr() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or or = new Or(EMPTY, equalsOf(fa, ONE), equalsOf(fa, TWO));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(In.class, e.getClass());
-        In in = (In) e;
-        assertEquals(fa, in.value());
-        assertThat(in.list(), contains(ONE, TWO));
-    }
-
-    public void testTwoEqualsWithSameValue() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or or = new Or(EMPTY, equalsOf(fa, ONE), equalsOf(fa, ONE));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(Equals.class, e.getClass());
-        Equals eq = (Equals) e;
-        assertEquals(fa, eq.left());
-        assertEquals(ONE, eq.right());
-    }
-
-    public void testOneEqualsOneIn() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or or = new Or(EMPTY, equalsOf(fa, ONE), new In(EMPTY, fa, List.of(TWO)));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(In.class, e.getClass());
-        In in = (In) e;
-        assertEquals(fa, in.value());
-        assertThat(in.list(), contains(ONE, TWO));
-    }
-
-    public void testOneEqualsOneInWithSameValue() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or or = new Or(EMPTY, equalsOf(fa, ONE), new In(EMPTY, fa, asList(ONE, TWO)));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(In.class, e.getClass());
-        In in = (In) e;
-        assertEquals(fa, in.value());
-        assertThat(in.list(), contains(ONE, TWO));
-    }
-
-    public void testSingleValueInToEquals() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Equals equals = equalsOf(fa, ONE);
-        Or or = new Or(EMPTY, equals, new In(EMPTY, fa, List.of(ONE)));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(equals, e);
-    }
-
-    public void testEqualsBehindAnd() {
-        FieldAttribute fa = getFieldAttribute();
-
-        And and = new And(EMPTY, equalsOf(fa, ONE), equalsOf(fa, TWO));
-        Filter dummy = new Filter(EMPTY, relation(), and);
-        LogicalPlan transformed = new CombineDisjunctionsToIn().apply(dummy);
-        assertSame(dummy, transformed);
-        assertEquals(and, ((Filter) transformed).condition());
-    }
-
-    public void testTwoEqualsDifferentFields() {
-        FieldAttribute fieldOne = getFieldAttribute("ONE");
-        FieldAttribute fieldTwo = getFieldAttribute("TWO");
-
-        Or or = new Or(EMPTY, equalsOf(fieldOne, ONE), equalsOf(fieldTwo, TWO));
-        Expression e = new CombineDisjunctionsToIn().rule(or);
-        assertEquals(or, e);
-    }
-
-    public void testMultipleIn() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or firstOr = new Or(EMPTY, new In(EMPTY, fa, List.of(ONE)), new In(EMPTY, fa, List.of(TWO)));
-        Or secondOr = new Or(EMPTY, firstOr, new In(EMPTY, fa, List.of(THREE)));
-        Expression e = new CombineDisjunctionsToIn().rule(secondOr);
-        assertEquals(In.class, e.getClass());
-        In in = (In) e;
-        assertEquals(fa, in.value());
-        assertThat(in.list(), contains(ONE, TWO, THREE));
-    }
-
-    public void testOrWithNonCombinableExpressions() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or firstOr = new Or(EMPTY, new In(EMPTY, fa, List.of(ONE)), lessThanOf(fa, TWO));
-        Or secondOr = new Or(EMPTY, firstOr, new In(EMPTY, fa, List.of(THREE)));
-        Expression e = new CombineDisjunctionsToIn().rule(secondOr);
-        assertEquals(Or.class, e.getClass());
-        Or or = (Or) e;
-        assertEquals(or.left(), firstOr.right());
-        assertEquals(In.class, or.right().getClass());
-        In in = (In) or.right();
-        assertEquals(fa, in.value());
-        assertThat(in.list(), contains(ONE, THREE));
-    }
-
     // Test BooleanFunctionEqualsElimination
     public void testBoolEqualsSimplificationOnExpressions() {
         BooleanFunctionEqualsElimination s = new BooleanFunctionEqualsElimination();
